Move vox2mm to NVUtilities, convert jest test to unit test
